The commonalities execution tests don't work when run as Junit plugin tests from within Eclipse. They only properly work when run during the maven tycho build.
One part of the issue is that the ExternalReactionsGenerator dynamically sets up a guice injector with the ReactionsLanguageRuntimeModule, which is missing the Eclipse specific overrides from the module in the UI project. This is resolved by setting up the UI injector of the Reactions Language at the beginning of the execution test via the new CombinedUiInjectorProvder. With this setup, the test cases work in both Eclipse and when run as maven tycho build.

The second part of the issue is that Xtext is not able to find the jvm types for the generated intermediate model(s) when serializing the generated reactions. This seems to only affect the generation of the field at the top of the generated reaction class. This is currently worked around by letting AccessibleElement create new, unassociated jvm types in memory with the correct qualified names, if it encounters an unknown jvm type reference. The proper solution would be to get Xtext to find the generated intermediate model types.

These changes also resolve the need for triggering two subsequent Xtext builds inside the execution test. However, the mentioned workaround does actually not work correctly when triggering another, second Xtext build afterwards.